requirements.txt
django-cors-headers>=4.0
Django>=4.2
django-environ>=0.10
Pillow>=10.0
numpy>=1.26
opencv-python>=4.8
rembg>=2.0
celery>=5.3
requests>=2.32
